moved modal window to separate component
